+++ src/components/manage/Blocks/Block/BlocksForm.jsx
@@ -16,6 +16,7 @@ import EditBlockWrapper from './EditBlockWrapper';
 import { setSidebarTab } from '@plone/volto/actions';
 import { useDispatch } from 'react-redux';
 import { useDetectClickOutside } from '@plone/volto/helpers';
+import QuantaBlockWrapper from './QuantaBlockWrapper';
 import config from '@plone/volto/registry';
 
 const BlocksForm = (props) => {
@@ -142,11 +143,21 @@ const BlocksForm = (props) => {
     onChangeFormData(newFormData);
   };
 
-  const defaultBlockWrapper = ({ draginfo }, editBlock, blockProps) => (
-    <EditBlockWrapper draginfo={draginfo} blockProps={blockProps}>
-      {editBlock}
-    </EditBlockWrapper>
-  );
+  const defaultBlockWrapper = ({ draginfo }, editBlock, blockProps) => {
+    if (config.settings.useQuantaToolbar) {
+      return (
+        <QuantaBlockWrapper draginfo={draginfo} blockProps={blockProps}>
+          {editBlock}
+        </QuantaBlockWrapper>
+      );
+    } else {
+      return (
+        <EditBlockWrapper draginfo={draginfo} blockProps={blockProps}>
+          {editBlock}
+        </EditBlockWrapper>
+      );
+    }
+  };
 
   const editBlockWrapper = children || defaultBlockWrapper;