+++ src/components/manage/Blocks/Text/Edit.jsx
@@ -5,8 +5,6 @@
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Button } from 'semantic-ui-react';
-import { doesNodeContainClick } from 'semantic-ui-react/dist/commonjs/lib';
 import Editor from 'draft-js-plugins-editor';
 import { convertFromRaw, convertToRaw, EditorState, RichUtils } from 'draft-js';
 import createInlineToolbarPlugin from 'draft-js-inline-toolbar-plugin';
@@ -15,10 +13,12 @@ import isSoftNewlineEvent from 'draft-js/lib/isSoftNewlineEvent';
 import { defineMessages, injectIntl } from 'react-intl';
 import { includes, isEqual } from 'lodash';
 import { filterEditorState } from 'draftjs-filters';
+import { MutateBlockButton } from '@plone/volto/components';
+import { Plug } from '@plone/volto/components/manage/Pluggable';
+
 import config from '@plone/volto/registry';
 
-import { Icon, BlockChooser } from '@plone/volto/components';
-import addSVG from '@plone/volto/icons/circle-plus.svg';
+const PassThrough = (props) => props.children;
 
 const messages = defineMessages({
   text: {
@@ -96,7 +96,6 @@ class Edit extends Component {
       this.state = {
         editorState,
         inlineToolbarPlugin,
-        addNewBlockOpened: false,
       };
     }
 
@@ -113,7 +112,6 @@ class Edit extends Component {
       // See https://github.com/draft-js-plugins/draft-js-plugins/issues/800
       setTimeout(this.node.focus, 0);
     }
-    document.addEventListener('mousedown', this.handleClickOutside, false);
   }
 
   /**
@@ -152,15 +150,6 @@ class Edit extends Component {
       !isEqual(this.props.data, nextProps.data) ||
       !isEqual(this.state.editorState, nextState.editorState)
     );
-  }
-
-  /**
-   * Component will unmount
-   * @method componentWillUnmount
-   * @returns {undefined}
-   */
-  componentWillUnmount() {
-    document.removeEventListener('mousedown', this.handleClickOutside, false);
   }
 
   /**
@@ -205,22 +194,6 @@ class Edit extends Component {
     this.setState({ editorState });
   }
 
-  toggleAddNewBlock = (e) => {
-    e.preventDefault();
-    this.setState((state) => ({ addNewBlockOpened: !state.addNewBlockOpened }));
-  };
-
-  handleClickOutside = (e) => {
-    if (
-      this.props.blockNode.current &&
-      doesNodeContainClick(this.props.blockNode.current, e)
-    )
-      return;
-    this.setState(() => ({
-      addNewBlockOpened: false,
-    }));
-  };
-
   /**
    * Render method.
    * @method render
@@ -240,6 +213,7 @@ class Edit extends Component {
       this.props.data?.disableNewBlocks || this.props.detached;
     const { InlineToolbar } = this.state.inlineToolbarPlugin;
     const { settings } = config;
+    const PlugInsert = settings.useQuantaToolbar ? Plug : PassThrough;
 
     return (
       <>
@@ -316,33 +290,21 @@ class Edit extends Component {
             this.node = node;
           }}
         />
-        <InlineToolbar />
-        {this.props.selected &&
-          !disableNewBlocks &&
-          !config.blocks.blocksConfig[
-            this.props.data?.['@type'] || 'text'
-          ].blockHasValue(this.props.data) && (
-            <Button
-              basic
-              icon
-              onClick={this.toggleAddNewBlock}
-              className="block-add-button"
-            >
-              <Icon name={addSVG} className="block-add-button" size="24px" />
-            </Button>
-          )}
-        {this.state.addNewBlockOpened && (
-          <BlockChooser
+        <PlugInsert pluggable="block-toolbar" id="block-text-toolbar">
+          <InlineToolbar />
+        </PlugInsert>
+        {this.props.selected && !config.settings.useQuantaToolbar && (
+          <MutateBlockButton
+            data={this.props.data}
+            block={this.props.block}
             onInsertBlock={(id, value) => {
-              this.setState((state) => ({
-                addNewBlockOpened: !state.addNewBlockOpened,
-              }));
               this.props.onSelectBlock(this.props.onInsertBlock(id, value));
             }}
-            currentBlock={this.props.block}
+            onMutateBlock={() => {}}
             allowedBlocks={this.props.allowedBlocks}
-            showRestricted={this.props.showRestricted}
             blocksConfig={this.props.blocksConfig}
+            size="24px"
+            className="block-add-button"
           />
         )}
       </>