+++ src/components/theme/App/App.jsx
@@ -110,7 +110,6 @@ class App extends Component {
     return (
       <PluggablesProvider>
         <BodyClass className={`view-${action}view`} />
-
         {/* Body class depending on content type */}
         {this.props.content && this.props.content['@type'] && (
           <BodyClass
@@ -119,7 +118,6 @@ class App extends Component {
               .toLowerCase()}`}
           />
         )}
-
         {/* Body class depending on sections */}
         <BodyClass
           className={cx({