+++ src/config/index.js
@@ -153,6 +153,7 @@ let config = {
       // },
     ],
     showSelfRegistration: false,
+    useQuantaToolbar: true,
     contentMetadataTagsImageField: 'image',
     hasWorkingCopySupport: false,
   },